fix: add_tagset duplicate events
The add_tagset component on the Post page was creating duplicate
events do to the setup function being re-executed every time an HTMX
swap happened. It was initially implemented that way because I didn't
realize the edit_post endpoint was missing the "post" context required
by the add_tagset component to properly configure the hx-post attribute
of the confirm button.