perf: add post thumbnail generation
Post thumbnails are generated on save. For now just for Images, but
in the future for video as well. Audios will probably just use a default
thumbnail image since there's nothing meaningful to extract for a
thumbnail. Currently defaults to a minimum height and width of 400px.

Also updates files to include thumbnails for demo.json fixture